+++ tests/api_clients/async_test_client/schemas.py
@@ -7,6 +7,8 @@ import typing
 
 import pydantic
 
+from clientele.schemas import ListResponse  # noqa
+
 
 class AnotherModel(pydantic.BaseModel):
     key: str